+++ overlays/nix/flake.nix
@@ -6,7 +6,7 @@
 
   outputs = { self, nixpkgs, nix, flake-utils }: {
     overlay = final: prev: {
-      nix = nix.defaultPackage.${prev.system}.overrideAttrs (old: {
+      nix = nix.packages.${prev.system}.default.overrideAttrs (old: {
         # WORKAROUND: When overriding Nix on Darwin, Nix throws a disallowed reference to boost error
         #             https://github.com/NixOS/nix/pull/5915
         disallowedReferences = if prev.stdenv.hostPlatform.isDarwin then [ ] else old.disallowedReferences;
@@ -17,7 +17,7 @@
           (prev.fetchpatch {
             name = "add-dirtyRev-and-dirtyShortRev-to-fetchGit.patch";
             url = "https://patch-diff.githubusercontent.com/raw/NixOS/nix/pull/5385.patch";
-            sha256 = "sha256-8PpWo31IN6T1GCRXF0BmtHRNpF6gyQKmz4kp2QdNa2c=";
+            sha256 = "sha256-zDnnty9ScICGB/H9EQsbjCFvp70Fbpa0+UnxouxTUxI=";
           })
         ];
       });
@@ -26,7 +26,7 @@
     flake-utils.lib.eachDefaultSystem (system: {
       packages.nix = (import nixpkgs { inherit system; overlays = [ self.overlay ]; }).nix;
 
-      defaultPackage = self.packages.${system}.nix;
+      packages.default = self.packages.${system}.nix;
     })
   );
 }